basic support for output parameters in jobs
In local scheduler mode only, for now.
More info is stored in the database (output json files, parameters links...)
Jobs are "free to" write a json file named via the SOMAWF_OUTPUT_PARAMS env variable. This file, if written, is read when a job ends successfully, and later jobs parameters are updated accordingly.
The GUI is not updated yet.