fixing version number
The version number has been messed-up at some point. The master branch
used to be 3.3 and now is has been "rewinded" to 3.2, which is the same
versions as in the brainvisa-5.1 branch.
See 2023/12/19
https://github.com/populse/soma-workflow/commit/3e48e9d294cc9b80a837f72617f7878467c22c5d
wheras one year earlier, 2022/12/22
https://github.com/populse/soma-workflow/commit/0c58fa67649680ac9e01bb9be73494fa29eaf321
I don't actually know if the latest pip releases have been made from the
branch brainvisa-5.1 or from master...